refactor(test): rename `kubernetes-module` projects to `kubernetes-type`
