+++ core/test/data/test-projects/kubernetes-type/api-image/garden.yml
