+++ core/test/data/test-projects/kubernetes-type/api-image/requirements.txt
