+++ core/test/data/test-projects/kubernetes-type/artifacts/garden.yml
