+++ core/test/data/test-projects/kubernetes-type/config-map-list/garden.yml
