+++ core/test/data/test-projects/kubernetes-type/config-map-list/manifests.yaml
