+++ core/test/data/test-projects/kubernetes-type/module-simple/garden.yml
