+++ core/test/data/test-projects/kubernetes-type/with-namespace/garden.yml
