+++ core/test/data/test-projects/kubernetes-type/with-source-module/garden.yml
