+++ core/test/integ/src/plugins/kubernetes/kubernetes-type/config.ts
