+++ core/test/integ/src/plugins/kubernetes/kubernetes-type/kubernetes-exec-test.ts
@@ -11,10 +11,10 @@ import { expect } from "chai"
 import { TestGarden, findNamespaceStatusEvent } from "../../../../../helpers"
 import { ConfigGraph } from "../../../../../../src/graph/config-graph"
 import { getKubernetesTestGarden } from "./common"
-import { RunTask } from "../../../../../../src/tasks/run"
 import { clearRunResult } from "../../../../../../src/plugins/kubernetes/run-results"
+import { TestTask } from "../../../../../../src/tasks/test"
 
-describe("kubernetes-type exec Run", () => {
+describe("kubernetes-type exec Test", () => {
   let garden: TestGarden
   let graph: ConfigGraph
 
@@ -26,10 +26,10 @@ describe("kubernetes-type exec Run", () => {
     graph = await garden.getConfigGraph({ log: garden.log, emit: false })
   })
 
-  it("should run a basic Run", async () => {
-    const action = graph.getRun("echo-run-exec")
+  it("should run a basic Test", async () => {
+    const action = graph.getTest("echo-test-exec")
 
-    const testTask = new RunTask({
+    const testTask = new TestTask({
       garden,
       graph,
       action,
@@ -45,8 +45,8 @@ describe("kubernetes-type exec Run", () => {
 
     garden.events.eventLog = []
     const results = await garden.processTasks({ tasks: [testTask], throwOnError: true })
+    expect(findNamespaceStatusEvent(garden.events.eventLog, "kubernetes-type-test-default")).to.exist
     const result = results.results.getResult(testTask)
-    expect(findNamespaceStatusEvent(garden.events.eventLog, "kubernetes-module-test-default")).to.exist
 
     expect(result).to.exist
     expect(result?.result).to.exist
