+++ core/test/integ/src/plugins/kubernetes/kubernetes-type/kubernetes-pod-run.ts
@@ -50,7 +50,7 @@ describe("kubernetes-type pod Run", () => {
     garden.events.eventLog = []
     const results = await garden.processTasks({ tasks: [testTask], throwOnError: true })
     const result = results.results.getResult(testTask)
-    expect(findNamespaceStatusEvent(garden.events.eventLog, "kubernetes-module-test-default")).to.exist
+    expect(findNamespaceStatusEvent(garden.events.eventLog, "kubernetes-type-test-default")).to.exist
 
     expect(result).to.exist
     expect(result?.result).to.exist
